requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
backoff==2.2.1
bcrypt==4.2.0
blinker==1.8.2
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.5.16
click==8.1.7
coloredlogs==15.0.1
dataclasses-json==0.6.7
Deprecated==1.2.14
distro==1.9.0
durationpy==0.9
exceptiongroup==1.2.2
fastapi==0.115.4
filelock==3.16.1
flatbuffers==24.3.25
frozenlist==1.5.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.35.0
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.67.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.4.0
importlib_resources==6.4.5
Jinja2==3.1.4
jiter==0.6.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kubernetes==31.0.0
langchain==0.3.4
langchain-chroma==0.1.4
langchain-community==0.3.3
langchain-core==0.3.13
langchain-openai==0.2.4
langchain-text-splitters==0.3.0
langgraph==0.2.39
langgraph-checkpoint==2.0.2
langgraph-sdk==0.1.35
langsmith==0.1.137
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.0
mdurl==0.1.2
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.12.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.19.2
openai==1.52.2
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.10
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pillow==10.4.0
posthog==3.7.0
propcache==0.2.0
protobuf==4.25.5
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.3
rpds-py==0.20.0
rsa==4.9
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
streamlit==1.39.0
sympy==1.13.3
tenacity==9.0.0
tiktoken==0.8.0
tokenizers==0.20.1
toml==0.10.2
tomli==2.0.2
tornado==6.4.1
tqdm==4.66.6
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
watchdog==5.0.3
watchfiles==0.24.0
websocket-client==1.8.0
websockets==13.1
wrapt==1.16.0
yarl==1.17.0
zipp==3.20.2
